breaking out ModuleSet
